# chatgpt-plugin-config/openapi.yaml
openapi: 3.0.0
info:
  title: APISIX Admin API
  description: >-
    APISIX Admin API is a RESTful API that allows you to create and manage
    APISIX resources.
  version: 3.3.0
servers:
  - url: http://localhost:5000
    description: Dev Environment  
...
paths:
...
  /apisix/admin/plugin_configs/{id}:  
    get:
      operationId: getPluginConfigById
      summary: Get Plugin Config by ID
      deprecated: false
      description: Get a plugin config by ID.
      tags:
        - Plugin Config
      parameters:
        - name: id
          in: path
          description: Specified plugin config ID.
          required: true
          example: ''
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    properties:
                      plugins:
                        type: object
                        properties: {}
                      desc:
                        maxLength: 256
                        type: string
                      labels:
                        type: object
                        patternProperties:
                          .*:
                            minLength: 1
                            maxLength: 64
                            pattern: ^\S+$
                            type: string
                            description: value of label
                        description: Key-value pairs to specify attributes.
                        properties: {}
                      id:
                        anyOf:
                          - type: string
                            minLength: 1
                            maxLength: 64
                            pattern: ^[a-zA-Z0-9-_.]+$
                          - minimum: 1
                            type: integer
                      update_time:
                        type: integer
                      create_time:
                        type: integer
                    type: object
                    required:
                      - id
                      - plugins
                  key:
                    type: string
                  modifiedIndex:
                    type: integer
                  createdIndex:
                    type: integer
              examples:
                '1':
                  summary: Successful Example
                  value:
                    value:
                      plugins:
                        limit-count:
                          rejected_code: 503
                          time_window: 60
                          key: remote_addr
                          key_type: var
                          count: 2
                          allow_degradation: false
                          show_limit_quota_header: true
                          policy: local
                      desc: desc
                      id: '1'
                      create_time: 1684380007
                      update_time: 1684380034
                    key: /apisix/plugin_configs/1
                    modifiedIndex: 151
                    createdIndex: 150
        '404':
          description: Key not found
          content: {}
  /apisix/admin/plugin_configs/{patch-id}:   
    patch:
      operationId: updatePluginConfigById
      summary: Update Plugin Config
      deprecated: false
      description: Update a plugin config.
      tags:
        - Plugin Config
      parameters:
        - name: patch-id
          in: path
          description: Specified plugin config ID.
          required: true
          example: ''
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                plugins:
                  type: object
                  properties: {}
                desc:
                  maxLength: 256
                  type: string
                labels:
                  type: object
                  patternProperties:
                    .*:
                      minLength: 1
                      maxLength: 64
                      pattern: ^\S+$
                      type: string
                      description: value of label
                  description: Key-value pairs to specify attributes.
                  properties: {}
                id:
                  anyOf:
                    - type: string
                      minLength: 1
                      maxLength: 64
                      pattern: ^[a-zA-Z0-9-_.]+$
                    - minimum: 1
                      type: integer
                update_time:
                  type: integer
                create_time:
                  type: integer
              type: object
              required:
                - id
                - plugins
            example:
              plugins:
                prometheus: {}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  value:
                    properties:
                      plugins:
                        type: object
                        properties: {}
                      desc:
                        maxLength: 256
                        type: string
                      labels:
                        type: object
                        patternProperties:
                          .*:
                            minLength: 1
                            maxLength: 64
                            pattern: ^\S+$
                            type: string
                            description: value of label
                        description: Key-value pairs to specify attributes.
                        properties: {}
                      id:
                        anyOf:
                          - type: string
                            minLength: 1
                            maxLength: 64
                            pattern: ^[a-zA-Z0-9-_.]+$
                          - minimum: 1
                            type: integer
                      update_time:
                        type: integer
                      create_time:
                        type: integer
                    type: object
                    required:
                      - id
                      - plugins
              examples:
                '1':
                  summary: Successful Example
                  value:
                    key: /apisix/plugin_configs/1
                    value:
                      plugins:
                        limit-count:
                          rejected_code: 503
                          time_window: 60
                          key: remote_addr
                          key_type: var
                          count: 2
                          allow_degradation: false
                          show_limit_quota_header: true
                          policy: local
                        prometheus:
                          prefer_name: false
                      desc: desc
                      id: '1'
                      create_time: 1684380007
                      update_time: 1684380472
        '400':
          description: Invalid configuration
          content: {}
        '404':
          description: Key not found
          content: {}
    delete:
      operationId: deletePluginConfigById
      summary: Delete Plugin Config by ID
      deprecated: false
      description: Delete a plugin config by ID.
      tags:
        - Plugin Config
      parameters:
        - name: id
          in: path
          description: Specified plugin config ID.
          required: true
          example: ''
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  deleted:
                    type: string
              examples:
                '1':
                  summary: Successful Example
                  value:
                    key: /apisix/plugin_configs/1
                    deleted: '1'
        '404':
          description: Key not found
          content: {}